Update update_bag view to not handle removing items and reflect input
View used to be able to remove products from bag if a quantity input
selector was set to 0. In a later commit the product quantity will
update on change instead of clicking the update button so I have
removed this functionality from the view to stop accidental removal of
product. The quantity input selector can't be set to 0.