Update success payment webhook handler to handler successfully orders
This commit updates the susccess payment webhook handler to
process orders when both the checkout form has been submitted
and when the user has left the page whilst Stripe is handling the
payment and before the checkout success page has been rendered but the
payment has successfully been charged by Stripe.

This commit will also stop multiple instances of the order being
created by the webhook if it is already in the database.